refactor(web): name playground behavior controller
